Add TickableTask (#2216)
* Add autopush functionality to MuTE (#1976)

* Working auto push

* Revert wildcard import

* Addresssed reviews

* Fix reference issue

* MuTE Upgrade Casings (#1988)

* Added cleanroom upgrade casing

* Added inventory and tank upgrades

* Added tooltips to mute casings

* Added power upgrades

* Set player UUID when placing MuTE

* Add a Generic Processing Logic and extract methods (#1992)

* add a generic processing logic

* calculate tier in another method

* calculate power logic in another method

* MuTE inventory upgrade logic (#2082)

* Catch potential NPE

* Don't load name when it doesn't exist

* Potentially cause weird non-replicatable issue where registry ends up with different key

* Use proper block removal method

* Validate index before using it

* Don't open controller GUI from inventory upgrade

* semi-working concept

* sync the ID of the inventory upgrade to correctly remove it later

* remove unneeded boolean

---------


* MuTE casing structure element (#2105)

* Added class containing MuTE relevant structure elements

* Migrate MuTE to new structure element

* Formatting fix

* Use int array instead of int hashmap, since its expected these arrays will never get long enough to be faster as hashmap

* Delete old code

* Refactor Item and Fluid to be in separate logic classes (#2178)

* basics of inventory logic

* mostly working item logic

* working nbt saving/loading

* fluid handler

* FluidSlotHandler WIP

* fluid handler mostly working

* remove fluid handler from gt5u

* prepare for conversion

* use correct imports

* spotless

* more controller logic

* spotless

* final refactor. migration next

* spotless

* add more methods to logic classes

* convert almost everything to use new Logic

* spotless

* make mute casing mode an int

* allow pump cover to work with FluidInventoryLogic

* pumps work

* spotless

* make item inventory logic work with every item input thing

* rework Fluid Inventory Logic to work with all fluid inputs

* spotless

* address annotation reviews

* finish off todos

* missed to dos

* cleanup

Coke oven will get a new GUI when i get to it

* address review

* prevent npes from ControllerXXXLogic

* null checks

* Base work

* PollutionTask

* move package

* Fix generics

* Internal -> OverrideOnly

* rebase fix

* Ducttape addPollution

---------

+++ src/main/java/gregtech/api/logic/ComplexParallelProcessingLogic.java
@@ -24,7 +24,7 @@ public class ComplexParallelProcessingLogic {
     protected final FluidStack[][] outputFluids;
     protected final long[] availableEut;
     protected final long[] eut;
-    protected final long[] durations;
+    protected final int[] durations;
     protected final boolean[] isItemVoidProtected;
     protected final boolean[] isFluidVoidProtected;
     protected boolean isCleanroom;
@@ -42,7 +42,7 @@ public class ComplexParallelProcessingLogic {
         outputFluids = new FluidStack[maxComplexParallels][];
         eut = new long[maxComplexParallels];
         availableEut = new long[maxComplexParallels];
-        durations = new long[maxComplexParallels];
+        durations = new int[maxComplexParallels];
         isItemVoidProtected = new boolean[maxComplexParallels];
         isFluidVoidProtected = new boolean[maxComplexParallels];
     }
@@ -167,7 +167,7 @@ public class ComplexParallelProcessingLogic {
         return true;
     }
 
-    public long getDuration(int index) {
+    public int getDuration(int index) {
         if (index >= 0 && index < maxComplexParallels) {
             return durations[index];
         }
